修复搜索

+++ src/main/java/com/gto/gtocore/api/machine/mana/feature/IManaMultiblock.java
@@ -10,22 +10,22 @@ public interface IManaMultiblock {
 
     boolean isGeneratorMana();
 
-    default int addMana(int amount, int limit) {
-        int change = 0;
+    default long addMana(long amount, int limit) {
+        long change = 0;
         for (IManaContainer container : getManaContainer()) {
             if (amount <= 0) return change;
-            int mana = container.addMana(amount, limit);
+            long mana = container.addMana(amount, limit);
             change += mana;
             amount -= mana;
         }
         return change;
     }
 
-    default int removeMana(int amount, int limit) {
-        int change = 0;
+    default long removeMana(long amount, int limit) {
+        long change = 0;
         for (IManaContainer container : getManaContainer()) {
             if (amount <= 0) return change;
-            int mana = container.removeMana(amount, limit);
+            long mana = container.removeMana(amount, limit);
             change += mana;
             amount -= mana;
         }
